add time_periods type constraint

+++ src/device_models/electric_loads.jl
@@ -1,4 +1,4 @@
-function loadvariables(m::JuMP.Model, devices_netinjection:: A, devices::Array{T,1}, time_periods) where {A <: PowerExpressionArray, T <: PowerSystems.ElectricLoad}
+function loadvariables(m::JuMP.Model, devices_netinjection:: A, devices::Array{T,1}, time_periods::Int64) where {A <: PowerExpressionArray, T <: PowerSystems.ElectricLoad}
     on_set = [d.name for d in devices if (d.available == true && !isa(d,PowerSystems.StaticLoad))]
 
     t = 1:time_periods
